requirements.txt
fastapi
uvicorn[standard]
python-multipart
qdrant-client==1.9.2
sentence-transformers
torch
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
pydantic
requests
python-dotenv
numpy
sqlmodel
python-docx
pypdf